tests/wpt/web-platform-tests/navigation-timing/test_performance_attributes_exist.html

<!DOCTYPE html>
<html>
    <head>
        <meta charset="utf-8" />
        <title>window.performance attributes</title>
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
        <link rel="help" href="http://www.w3.org/TR/navigation-timing/#sec-window.performance-attribute"/>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="/common/performance-timeline-utils.js"></script>
        <script src="resources/webperftestharness.js"></script>
    </head>
    <body>
        <h1>Description</h1>
        <p>This test validates that the navigation and timing attributes exist for window.performance
        (but does not validate that their values are correct).</p>
        <div id="log"></div>
        <script>
            test_namespace('navigation');
            test_namespace('timing', true);
        </script>
    </body>
</html>
